% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHQLImpute.R
\name{EHQLImpute}
\alias{EHQLImpute}
\title{Imputation of nonignorable missing data using the EHQL Gaussian Copula}
\usage{
EHQLImpute(
  YR,
  ncolY,
  ncolR,
  aux_quantiles,
  aux_infos,
  MA = rep(T, ncolY),
  nImps,
  nsamp,
  burn
)
}
\arguments{
\item{YR}{Observed data matrix. It should be partitioned into the study variables Y followed by missingness indicators R.}

\item{ncolY}{Number of study variables}

\item{ncolR}{Number of study variables modeled as noningorable}

\item{aux_quantiles}{List of length ncolY containing auxiliary quantiles assumed known for each study variable. aux_quantiles[[j]] should be a vector containg 0, intermediate quantiles, and 1.}

\item{aux_infos}{List of length ncolY containing the known auxiliary values of the study variables. aux_infos[[j]] should be a vector containing \eqn{(F_j^{-1}(0),\dots,F_{j}^{-1}(\tau),\dots,F_{j}^{-1}(1))}. In addition aux_infos[[j]] should be the same length as aux_quantiles[[j]]}

\item{MA}{Vector logicals of length ncolY indicating whether or not to estimate the margin adjustment at intermediate quantiles using the margin adjustment. Default is True for each study variable, which specifies an evenly spaced
grid of 15 points across \eqn{Y_j^{obs}}, and learns the intermediate quantiles of these points during the EHQL Gibbs sampler. These posterior samples are then smoothed and used for imputation.}

\item{nImps}{Number of completed data sets to produce}

\item{nsamp}{Number of iterations for the MCMC}

\item{burn}{Number of Burn-in samples for the MCMC}
}
\value{
C.post,alpha.post, and MAs, which are the nsamp-burn posterior samples of the copula correlation, latent intercept vector, and margin adjustment (list or length ncolY, MAs[[j]] computes the margin adjustment at x_ma[[j]]), respectively. In addition, YImpute is a list of length nImps containing imputed data sets.
}
\description{
The user specifies auxiliary quantiles of the marginal distribution of each study variable,
and inputs an observed data matrix (Y,R) containing the observed and missing study variables (Y) and missingness indicators.
The function estimates the EHQL Gaussian copula, returns posterior samples of the copula parameters, and creates nImps completed data sets
}
