% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_Z_MA.R
\name{resample_Z_MA}
\alias{resample_Z_MA}
\title{Title}
\usage{
resample_Z_MA(
  j,
  Z,
  Y,
  R,
  Rlevels,
  alpha,
  Lambda,
  eta,
  Sigma.diag,
  plugin.marginal,
  is_cat_bin,
  have_aux,
  aux_quantiles,
  aux_bins,
  Z_past
)
}
\arguments{
\item{j}{s}

\item{Z}{s}

\item{Y}{s}

\item{R}{s}

\item{Rlevels}{s}

\item{alpha}{s}

\item{Lambda}{s}

\item{eta}{s}

\item{Sigma.diag}{s}

\item{plugin.marginal}{s}

\item{is_cat_bin}{s}

\item{have_aux}{s}

\item{aux_quantiles}{s}

\item{aux_bins}{s}

\item{Z_past}{s}
}
\value{
s
}
\description{
Title
}
