% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMGP.R
\name{sampleMGP}
\alias{sampleMGP}
\title{Title}
\usage{
sampleMGP(theta.jh, delta.h, a1 = 3, a2 = 5)
}
\arguments{
\item{theta.jh}{sss}

\item{delta.h}{sss}

\item{a1}{sss}

\item{a2}{sss}
}
\value{
ss
}
\description{
Title
}
