% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_Z_MA.R
\name{resample_Z_MA}
\alias{resample_Z_MA}
\title{Sample latent variables under the EHQL Gaussian copula}
\usage{
resample_Z_MA(
  j,
  Z,
  Y,
  R,
  Rlevels,
  alpha,
  Lambda,
  eta,
  Sigma.diag,
  plugin.marginal,
  is_cat_bin,
  have_aux,
  aux_quantiles,
  aux_bins,
  Z_past
)
}
\arguments{
\item{j}{Variable Index}

\item{Z}{Current iteration of the latent data matrix Z}

\item{Y}{Observed data matrix}

\item{R}{nx2p matrix of ordered quantile bin memberships for Y and levels for R}

\item{Rlevels}{Vector of length \eqn{2 \times p} with the number of unique auxiliary bins + intermediate quantiles}

\item{alpha}{Vector of length \eqn{2\times p} with intercept terms for unordered categorical variables/ components of R}

\item{Lambda}{Current iteration of the latent factor loadings}

\item{eta}{Current iteration of nxk.star latent factor matrix}

\item{Sigma.diag}{Current iteration of diagonal error variances}

\item{plugin.marginal}{Vector of length \eqn{2 \times p} indicating whether EHQL sampling should take place. By default should be set to F in each component}

\item{is_cat_bin}{Vector of length \eqn{2\times p} indicating which components of the data matrix YR correspond to R or unordered categorical variables}

\item{have_aux}{Vector of length p indicating whether we have auxiliary information on each study variable}

\item{aux_quantiles}{List of length p containing the auxiliary quantels available for each study variable j}

\item{aux_bins}{List of length p, each containing a vector of length nobs indicating which quantile interval \eqn{y_{ij}^{obs}} belongs to}

\item{Z_past}{past sample of latent Z -- useful for initial stages of MCMC to ensure EHQL orderings are consistent}
}
\value{
nx2p posterior sample of Z
}
\description{
Iterates over the study variables and missingness indicators and samples them according to EHQL ordering constraints (for Y) and probit data augmentation for R
}
