% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_Z_MA.R
\name{resample_Z_MA}
\alias{resample_Z_MA}
\title{Title}
\usage{
resample_Z_MA(
  j,
  Z,
  Y,
  R,
  Rlevels,
  alpha,
  Lambda,
  eta,
  Sigma.diag,
  plugin.marginal,
  is_cat_bin,
  have_aux,
  aux_quantiles,
  aux_bins,
  Z_past
)
}
\description{
Title
}
